Update patch set 1
Patch Set 1: Code-Review+1

Yo,

Thanks for the review, LGTM. I am gonna +1 it until monday.

With this review, I've realized that we didn't gate this project, so basically a change to this project can break all our dci-rpm-build jobs down the line.

So instead of beeing a noop job, I've created a job that will trigger a build for all the packages we use. As you can see this patch hasn't broken anything as it got a +1.

Now, why monday? It is currently testing all the rpm's but dci-control-server. This is due to the fact of us using weird naming convention for no reason.

So my goal is on monday morning to merge all the reviews related with that[1] and then merge this one. And I am not doing it today because its friday :)

[1] https://softwarefactory-project.io/r/#/q/status:open+branch:master+topic:rename_dci_to_dci_control_server

Patch-set: 1
Label: Code-Review=+1