Fix issue with centos-sclo-rh & rh-nodejs10
In a recent update, mock templates now include the centos-sclo-rh repo
with a whitelist "includepkgs=devtoolset*".

For dci-ui we need to install rh-nodejs10 from the centos-sclo-rh repo,
which dci-packaging adds to yum's configuration. However this 2nd
declaration is now ignored due to the first declaration in the mock
template.

This commit adds renames the scl repo used by DCI to
"centos-sclo-rh-for-dci" to allow installing required packages.

Change-Id: I693af338efdecd423566d0302217f55e61cd7997